fix(hive): fix ARM format warning and update python-keepkey pin
- Cast account_index to (unsigned int) to suppress -Werror=format on ARM
  (uint32_t is long unsigned int on 32-bit targets; %u expects unsigned int)
- Pin deps/python-keepkey to 04119f3 (old-style pb2 compatible with
  protobuf 3.20.3 CI runtime)

+++ lib/firmware/fsm_msg_hive.h
@@ -106,7 +106,8 @@ void fsm_msgHiveGetPublicKeys(const HiveGetPublicKeys* msg) {
 
   if (msg->has_show_display && msg->show_display) {
     if (!confirm(ButtonRequestType_ButtonRequest_Other, "Hive Keys",
-                 "Export all Hive keys for account %u?", account_index)) {
+                 "Export all Hive keys for account %u?",
+                 (unsigned int)account_index)) {
       memzero(root, sizeof(*root));
       fsm_sendFailure(FailureType_Failure_ActionCancelled, _("Cancelled"));
       layoutHome();
